connected front end with rails api for backend

+++ db/migrate/20221111153258_create_characters.rb
@@ -2,8 +2,8 @@ class CreateCharacters < ActiveRecord::Migration[7.0]
   def change
     create_table :characters do |t|
       t.string :name
-      t.integer :x_location
-      t.integer :y_location
+      t.decimal :x_location
+      t.decimal :y_location
       t.string :picture
       t.references :image, null: false, foreign_key: true
